requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pymupdf
openai
